Add example programs
There was a program in the src/user directory which used the library
to send an OSC messages. This example wasn't very useful without a
receiver. I decided to drop it, and instead add examples for sending
and receiving messags. The send example (in example/send/send.pony) is
compatible with the type of message expected by the OSC_recv.ck
example code for the Chuck project. The receive example (in
example/receive/receive.pony) shows the address and arguments of
incoming OSC messages, and can be used with the send.pony example.